Fix android react-native link
