+++ android/src/main/AndroidManifest.xml
@@ -1,4 +1,3 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.i18n.reactnativei18n">
-
+    package="com.alexanderzaytsev.rn18n">
 </manifest>
